+++ resources/views/estimator/show.blade.php
@@ -11,7 +11,9 @@
 
                     <p>Name: </p>
                     <b>{{$quotation->name}}</b>
+                    @if($quotation->status == \App\Enums\QuotationStatus::NEW)
                     <x-primary-button style="display: flex"  type="button" onclick="window.location='{{ route('estimations.edit',['id' => $quotation->id]) }}'">Edit</x-primary-button>
+                    @endif
                 </div>
             </div>
 
@@ -19,7 +21,7 @@
         <br>
         <div class="bg-white overflow-hidden shadow-sm sm:rounded-lg">
             <div class="p-6 text-gray-900">
-                @if($quotation->status == 'done')
+                @if($quotation->status == \App\Enums\QuotationStatus::DONE)
                 <div style="display: flex; justify-content: space-between">
                     <p>Frontend estimation:</p>
                     <form method="POST" action="{{ route('xlsx.download') }}">
@@ -47,13 +49,13 @@
                     </form>
 
                 </div>
-                @elseif($quotation->status == 'in_progress')
+                @elseif($quotation->status == \App\Enums\QuotationStatus::IN_PROGRESS)
                     <div style="text-align: center;color: darkorange">
                         <p>Processing...</p>
 
                     </div>
                 @else
-                    <form method="POST" action="{{ route('estimations.send') }}">
+                    <form method="POST" action="{{ route('estimations.dispatch') }}">
                         @csrf
                         <input type="hidden" name="id" value={{$quotation->id}}>
                         <div style="text-align: center">
@@ -66,8 +68,28 @@
                 @endif
             </div>
 
+
+
         </div>
 
+    </div>
+    <div class="sm:max-w-xl mx-auto p-4 sm:p-6 lg:p-8">
+    <div class="bg-white overflow-hidden shadow-sm sm:rounded-lg">
+        <div class="p-6 text-gray-900">
+            <div style="display: flex; justify-content: space-between">
+                <p>Application description:</p>
+                <form method="POST" action="{{ route('docx.download') }}">
+                    @csrf
+                    <input type="hidden" name="id" value={{$quotation->id}}>
+                    <div style="text-align: center">
+                        <x-primary-button>Download .docx</x-primary-button>
+                    </div>
+
+                </form>
+            </div>
+        </div>
+
+    </div>
     </div>
     <div class="max-w-5xl mx-auto p-4 sm:p-6 lg:p-8">